Move offsets cronjob to its own odigospro controller
