+++ helm/odigos/values.yaml
@@ -470,6 +470,8 @@ clickhouseDestinationJsonType:
 # Enable automatic Go library offsets updates.
 # See https://docs.odigos.io/pipeline/golang/ebpf#go-auto-offsets for more details.
 # cron schedule for automatic Go offsets updates (e.g. "0 0 * * *" for daily at midnight). Set to empty string to disable.
+# Custom Offsets support is only available in Odigos pro tier.
 goAutoOffsetsCron: ''
 # mode for automatic Go offsets updates. Options include direct (pull from cloud server), image (pull from local image) and off (disabled).
+# Custom Offsets support is only available in Odigos pro tier.
 goAutoOffsetsMode: 'off'
